fix warnings
git-svn-id: http://skia.googlecode.com/svn/trunk@68 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkFloatingPoint.h
@@ -47,8 +47,8 @@
     #define sk_float_floor(x)       floorf(x)
     #define sk_float_ceil(x)        ceilf(x)
 #ifdef SK_BUILD_FOR_MAC
-    #define sk_float_acos(x)        acos(x)
-    #define sk_float_asin(x)        asin(x)
+    #define sk_float_acos(x)        static_cast<float>(acos(x))
+    #define sk_float_asin(x)        static_cast<float>(asin(x))
 #else
     #define sk_float_acos(x)        acosf(x)
     #define sk_float_asin(x)        asinf(x)
